+++ tests/asyncio/conftest.py
@@ -1,0 +1,1 @@
+from tests.asyncio.dummy_app import *
