+++ tests/asyncio/test_basic_usage.py
@@ -1,37 +1,10 @@
 import asyncio
 
 import aiohttp
-import pytest
 
 from pytest_bdd import given, when, then, scenarios, parsers
-from tests.asyncio.app import create_server, setup_and_teardown_flask_app, DummyApp
 
 scenarios("basic_usage.feature")
-
-
-@pytest.fixture
-def dummy_server_host():
-    return "localhost"
-
-
-@pytest.fixture
-def dummy_server_port():
-    return 10000
-
-
-@pytest.fixture
-def launch_dummy_server(dummy_server_host, dummy_server_port):
-    with setup_and_teardown_flask_app(create_server(), dummy_server_host, dummy_server_port):
-        yield
-
-
-@pytest.fixture
-async def launch_dummy_app(event_loop, launch_dummy_server, dummy_server_host, dummy_server_port):
-    app = DummyApp(dummy_server_host, dummy_server_port)
-    task = event_loop.create_task(app.run())
-    yield
-    task.cancel()
-    await asyncio.sleep(0)
 
 
 @given("i have launched app")
